requirements.txt
streamlit>=1.32.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.12.0
pydantic>=2.6.0
python-dotenv>=1.0.0
trafilatura>=1.8.0
html2text>=2024.2.26
google-generativeai>=0.5.0

# Optional
# anthropic>=0.20.0
